\name{cfb-package}
\alias{cfb-package}
\alias{cfb}
\docType{package}
\title{
Collection of College Football scrapers and cleaners
}
\description{
Currently the Package parses through NCAA football JSON files to extract play-by-play data.
}
\details{
\tabular{ll}{
Package: \tab cfb\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-25\cr
License: \tab MIT \cr
}
}
\author{
Tyler Hunt <tyler@psychoanalytix.com>
}
