% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNGET.R
\name{ESPNGET}
\alias{ESPNGET}
\title{Pulls in all schedule pages for all teams}
\usage{
ESPNGET(Year)
}
\description{
Using curl we pull all 254 teams schedules html at once. Use this function sparingly as is hits ESPN's servers in an intense way. Takes about 25 seconds
}

