% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPNResults.R
\name{ESPNResults}
\alias{ESPNResults}
\title{Scrape ESPN Game Results.}
\usage{
ESPNResults(Year = 2016, ESPNSource = NULL)
}
\description{
Using curl we pull all 254 teams schedules at once. Use this function sparingly as is hits ESPN's servers in an intense way. Takes about 25 seconds
}

