% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCAAboxscore.R
\name{NCAAboxscore}
\alias{NCAAboxscore}
\title{Scrape NCAA boxscore json files.}
\usage{
NCAAboxscore(url)
}
\examples{
NCAAboxscore("http://data.ncaa.com/jsonp/game/football/fbs/2016/10/15/utah-oregon-st/boxscore.json")

}

