% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCAAscoringSummary.R
\name{NCAAscoringSummary}
\alias{NCAAscoringSummary}
\title{Scrape NCAA scoringSummary}
\usage{
NCAAscoringSummary(url)
}
\examples{
NCAAscoringSummary("http://data.ncaa.com/jsonp/game/football/fbs/2016/10/15/utah-oregon-st/scoringSummary.json")

}

