% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Numerics.R
\name{Numerics}
\alias{Numerics}
\title{Numerics}
\usage{
Numerics(data)
}
\arguments{
\item{data}{dataframe with potential numerics to convert.}
}
\description{
Convert miss classed numerics vector from factor/character to numeric
}

